# api/handlers/race_handler.go
package handlers

import (
	"dnd-api/api"
	"dnd-api/api/requests"
	"dnd-api/api/responses"
	"dnd-api/db/models"
	"dnd-api/pkg/closer"
	"fmt"
	"github.com/jinzhu/gorm"
	"github.com/labstack/echo/v4"
	"github.com/labstack/gommon/random"
	"log"
	"net/http"
	"path/filepath"
)

type RaceHandler struct {
	server *api.Server
}

func NewRaceHandler(server *api.Server) *RaceHandler {
	return &RaceHandler{
		server: server,
	}
}

// List godoc
// @Summary List races
// @Description List races (paginated)
// @ID races-list
// @Tags Race Actions
// @Accept json
// @Produce json
// @Param search query string false "Search races by name"
// @Param page query int false "The page number"
// @Param page_size query int false "The numbers of items to return. Max 100"
// @Success 200 {object} responses.RacePaginatedResponse
// @Router /races [get]
func (h *RaceHandler) List(c echo.Context) error {
	var scopes []func(db *gorm.DB) *gorm.DB

	search := c.QueryParam("search")
	if search != "" {
		searchTerm := fmt.Sprintf("%%%s%%", search)
		scopes = append(scopes, func(db *gorm.DB) *gorm.DB {
			return db.Where("name LIKE ?", searchTerm)
		})
	}

	races, page, pageSize := h.server.Repos.Race.GetRaces(c, scopes)
	count := h.server.Repos.Race.CountRaces(scopes)

	res := responses.NewRacePaginatedResponse(races, count, page, pageSize)
	return responses.Response(c, http.StatusOK, res)
}

// Get godoc
// @Summary Get race by ID
// @Description Get race by ID
// @ID races-get
// @Tags Race Actions
// @Accept json
// @Produce json
// @Param id path string true "ID"
// @Success 200 {object} responses.RaceResponse
// @Failure 404 {object} responses.Error
// @Router /races/{id} [get]
func (h *RaceHandler) Get(c echo.Context) error {
	id := c.Param("id")

	race := h.server.Repos.Race.GetById(id)
	if race.ID == 0 {
		return responses.ErrorResponse(c, http.StatusNotFound, "Race not found")
	}

	res := responses.NewRaceResponse(race)
	return responses.Response(c, http.StatusOK, res)
}

// Create godoc
// @Summary Create race
// @Description Create race
// @ID races-create
// @Tags Race Actions
// @Accept json
// @Produce json
// @Param params body requests.CreateRaceRequest true "Race information"
// @Success 201 {object} responses.RaceResponse
// @Failure 400 {object} responses.Error
// @Failure 500 {object} responses.Error
// @Router /races [post]
func (h *RaceHandler) Create(c echo.Context) error {
	request := new(requests.CreateRaceRequest)
	if err := c.Bind(request); err != nil {
		return responses.ErrorResponse(c, http.StatusBadRequest, err.Error())
	}
	if err := c.Validate(request); err != nil {
		return responses.ErrorResponse(c, http.StatusBadRequest, "Required fields are empty or not valid: "+err.Error())
	}

	race := &models.Race{
		Name:             request.Name,
		ShortDescription: request.ShortDescription,
		CreatureType:     request.CreatureType,
		Size:             request.Size,
		BaseSpeed:        request.BaseSpeed,
	}
	err := h.server.Repos.Race.Create(race)
	if err != nil {
		return responses.ErrorResponse(c, http.StatusInternalServerError, "Something went wrong creating the race")
	}

	res := responses.NewRaceResponse(race)
	return responses.Response(c, http.StatusCreated, res)
}

// Update godoc
// @Summary Update race
// @Description Update race
// @ID races-update
// @Tags Race Actions
// @Accept json
// @Produce json
// @Param id path string true "Race ID"
// @Param params body requests.UpdateRaceRequest true "Race information"
// @Success 200 {object} responses.RaceResponse
// @Failure 400 {object} responses.Error
// @Failure 404 {object} responses.Error
// @Failure 500 {object} responses.Error
// @Router /races/{id} [put]
func (h *RaceHandler) Update(c echo.Context) error {
	id := c.Param("id")

	request := new(requests.UpdateRaceRequest)
	if err := c.Bind(request); err != nil {
		return responses.ErrorResponse(c, http.StatusBadRequest, err.Error())
	}
	if err := c.Validate(request); err != nil {
		return responses.ErrorResponse(c, http.StatusBadRequest, "Required fields are empty or not valid: "+err.Error())
	}

	race := h.server.Repos.Race.GetById(id)
	if race.ID == 0 {
		return responses.ErrorResponse(c, http.StatusNotFound, "Race not found")
	}

	race.Name = request.Name
	race.ShortDescription = request.ShortDescription
	race.CreatureType = request.CreatureType
	race.Size = request.Size
	race.BaseSpeed = request.BaseSpeed

	err := h.server.Repos.Race.Update(race)
	if err != nil {
		return responses.ErrorResponse(c, http.StatusInternalServerError, "Something went wrong updating the race")
	}

	res := responses.NewRaceResponse(race)
	return responses.Response(c, http.StatusOK, res)
}

// Delete godoc
// @Summary Delete race
// @Description Delete race
// @ID races-delete
// @Tags Race Actions
// @Accept json
// @Produce json
// @Param id path string true "Race ID"
// @Success 200 {object} responses.Data
// @Failure 404 {object} responses.Error
// @Failure 500 {object} responses.Error
// @Router /races/{id} [delete]
func (h *RaceHandler) Delete(c echo.Context) error {
	id := c.Param("id")

	race := h.server.Repos.Race.GetById(id)
	if race.ID == 0 {
		return responses.ErrorResponse(c, http.StatusNotFound, "Race not found")
	}

	if race.Logo.ID != 0 {
		// Delete from file store
		err := h.server.Dependencies.GetFileStore().Delete(fmt.Sprintf("%v/%v", race.Logo.FileLocation, race.Logo.Filename))
		if err != nil {
			log.Println("Error deleting race logo from file store: ", err.Error())
			return responses.ErrorResponse(c, http.StatusInternalServerError, "Something went wrong deleting the race logo from the file store")
		}

		// Delete from database
		err = h.server.Repos.File.Delete(race.Logo)
		if err != nil {
			log.Println("Error deleting race logo from database: ", err.Error())
			return responses.ErrorResponse(c, http.StatusInternalServerError, "Something went wrong deleting the race logo from the database")
		}
	}

	err := h.server.Repos.Race.Delete(race)
	if err != nil {
		return responses.ErrorResponse(c, http.StatusInternalServerError, "Something went wrong deleting the race")
	}

	return responses.MessageResponse(c, http.StatusOK, "Race deleted successfully")
}

// UploadLogo godoc
// @Summary Upload race logo
// @Description Upload race logo
// @ID races-upload-logo
// @Tags Race File Actions
// @Accept json
// @Produce json
// @Param id path string true "Race ID"
// @Success 200 {object} responses.Data
// @Failure 400 {object} responses.Error
// @Failure 404 {object} responses.Error
// @Failure 500 {object} responses.Error
// @Router /races/{id}/upload/logo [post]
func (h *RaceHandler) UploadLogo(c echo.Context) error {
	id := c.Param("id")

	race := h.server.Repos.Race.GetById(id)
	if race.ID == 0 {
		return responses.ErrorResponse(c, http.StatusNotFound, "Race not found")
	}

	// Check the file mimetype - We only want to accept images
	fileName, fileMimeType, err := h.server.Dependencies.GetFileService().ReadFileInfo(c)
	if err != nil {
		log.Printf("Error reading file info: %v", err)
		return responses.ErrorResponse(c, http.StatusBadRequest, "Unable to read file")
	}
	fileExtension := filepath.Ext(fileName)
	allowedFileExtensions := []string{"jpg", "jpeg", "png", "webp"}
	if !h.server.Dependencies.GetFileService().FileExtensionAllowed(fileExtension, allowedFileExtensions) {
		return responses.ErrorResponse(c, http.StatusBadRequest, "Invalid file type")
	}

	allowedMIMEs := []string{"image/jpeg", "image/png", "image/webp"}
	if !h.server.Dependencies.GetFileService().MIMETypeAllowed(fileMimeType, allowedMIMEs) {
		return responses.ErrorResponse(c, http.StatusBadRequest, "Invalid file content type")
	}

	// Get file from request
	file, err := c.FormFile("file")
	if err != nil {
		log.Printf("Error getting file from form: %v", err)
		return responses.ErrorResponse(c, http.StatusInternalServerError, "Error getting file from request")
	}

	// Change the file name
	newFilename := random.String(32)
	if fileExtension != "" {
		newFilename += fileExtension
	}
	file.Filename = newFilename

	// Upload file
	src, err := file.Open()
	if err != nil {
		log.Printf("Error opening file: %v", err)
		return responses.ErrorResponse(c, http.StatusInternalServerError, "Error opening file")
	}
	defer closer.Close(src)

	path := fmt.Sprintf("races/%v", race.ID)
	err = h.server.Dependencies.GetFileStore().Save(path, src, file.Filename)
	if err != nil {
		log.Printf("Error saving file: %v", err)
		return responses.ErrorResponse(c, http.StatusInternalServerError, "Error saving file")
	}

	// Create DB record
	logoFile := &models.File{
		Model:        models.FileModelRaceLogo,
		ModelId:      race.ID,
		Filename:     newFilename,
		FileLocation: path,
	}
	if err := h.server.Repos.File.Create(logoFile); err != nil {
		log.Printf("Error creating file record: %v", err)
		return responses.ErrorResponse(c, http.StatusInternalServerError, "Error creating file record")
	}

	return responses.MessageResponse(c, http.StatusOK, "File uploaded")
}