# docs/swagger.yaml
basePath: /
definitions:
...
  requests.CreateRaceRequest:
    properties:
      base_speed:
        example: 30
        type: integer
      creature_type:
        example: Humanoid
        maxLength: 200
        type: string
      name:
        example: Dragonborn
        maxLength: 200
        type: string
      short_description:
        example: The ancestors of dragonborn hatched from the eggs of chromatic and
          metallic dragons.
        type: string
      size:
        example: Medium (about 5-7 feet tall)
        maxLength: 200
        type: string
    required:
    - base_speed
    - creature_type
    - name
    - short_description
    - size
    type: object
  requests.CreateSubclassRequest:
    properties:
      class_id:
        type: integer
      name:
        maxLength: 200
        type: string
      short_description:
        type: string
    required:
    - class_id
    - name
    - short_description
    type: object
...
  requests.UpdateRaceRequest:
    properties:
      base_speed:
        example: 30
        type: integer
      creature_type:
        example: Humanoid
        maxLength: 200
        type: string
      name:
        example: Dragonborn
        maxLength: 200
        type: string
      short_description:
        example: The ancestors of dragonborn hatched from the eggs of chromatic and
          metallic dragons.
        type: string
      size:
        example: Medium (about 5-7 feet tall)
        maxLength: 200
        type: string
    required:
    - base_speed
    - creature_type
    - name
    - short_description
    - size
    type: object
...
  responses.Data:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  responses.Error:
    properties:
      code:
        type: integer
      error:
        type: string
    type: object
  responses.FileResponse:
    properties:
      created_at:
        type: string
      enterprise_uid:
        type: string
      file_location:
        type: string
      filename:
        type: string
      id:
        example: 1
        type: integer
      model:
        type: string
      model_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  responses.RacePaginatedResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/responses.RaceResponse'
        type: array
      meta:
        $ref: '#/definitions/responses.ResponseMeta'
    type: object
  responses.RaceResponse:
    properties:
      base_speed:
        type: integer
      creature_type:
        type: string
      id:
        type: integer
      logo:
        $ref: '#/definitions/responses.FileResponse'
      name:
        type: string
      short_description:
        type: string
      size:
        type: string
    type: object
...
  responses.ResponseMeta:
    properties:
      page:
        example: 3
        type: integer
      page_size:
        example: 25
        type: integer
      total_count:
        example: 300
        type: integer
    type: object
  responses.SubclassPaginatedResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/responses.SubclassResponse'
        type: array
      meta:
        $ref: '#/definitions/responses.ResponseMeta'
    type: object
  responses.SubclassResponse:
    properties:
      class_id:
        type: integer
      id:
        type: integer
      logo:
        $ref: '#/definitions/responses.FileResponse'
      name:
        type: string
      short_description:
        type: string
    type: object
info:
  contact:
    email: jbrokenshire0306@gmail.com
    name: Jared Brokenshire
  description: A management solution for Dungeons & Dragons character sheet information
  title: D&D Character Dashboard
  version: 0.0.1
paths:
...
  /races:
    get:
      consumes:
      - application/json
      description: List races (paginated)
      operationId: races-list
      parameters:
      - description: Search races by name
        in: query
        name: search
        type: string
      - description: The page number
        in: query
        name: page
        type: integer
      - description: The numbers of items to return. Max 100
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.RacePaginatedResponse'
      summary: List races
      tags:
      - Race Actions
    post:
      consumes:
      - application/json
      description: Create race
      operationId: races-create
      parameters:
      - description: Race information
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.CreateRaceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.RaceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Create race
      tags:
      - Race Actions
  /races/{id}:
    delete:
      consumes:
      - application/json
      description: Delete race
      operationId: races-delete
      parameters:
      - description: Race ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Data'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Delete race
      tags:
      - Race Actions
    get:
      consumes:
      - application/json
      description: Get race by ID
      operationId: races-get
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.RaceResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Get race by ID
      tags:
      - Race Actions
    put:
      consumes:
      - application/json
      description: Update race
      operationId: races-update
      parameters:
      - description: Race ID
        in: path
        name: id
        required: true
        type: string
      - description: Race information
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.UpdateRaceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.RaceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Update race
      tags:
      - Race Actions
  /races/{id}/upload/logo:
    post:
      consumes:
      - application/json
      description: Upload race logo
      operationId: races-upload-logo
      parameters:
      - description: Race ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Data'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Upload race logo
      tags:
      - Race File Actions
  /subclasses:
    post:
      consumes:
      - application/json
      description: Create subclass
      operationId: subclasses-create
      parameters:
      - description: Subclass information
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/requests.CreateSubclassRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.SubclassResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Create subclass
      tags:
      - Subclass Actions
  /subclasses/{classId}:
    get:
      consumes:
      - application/json
      description: List subclasses for class (paginated)
      operationId: subclasses-list
      parameters:
      - description: Class ID
        in: path
        name: classId
        required: true
        type: string
      - description: Search subclasses by name
        in: query
        name: search
        type: string
      - description: The page number
        in: query
        name: page
        type: integer
      - description: The numbers of items to return. Max 100
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.SubclassPaginatedResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.Error'
      summary: List subclasses
      tags:
      - Subclass Actions
  /subclasses/{classId}/{subclassId}/upload/logo:
    post:
      consumes:
      - application/json
      description: Upload subclass logo
      operationId: subclasses-upload-logo
      parameters:
      - description: Class ID
        in: path
        name: classId
        required: true
        type: string
      - description: Subclass ID
        in: path
        name: subclassId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Data'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Upload subclass logo
      tags:
      - Subclass File Actions
...
swagger: "2.0"